Added Events
- added delete
- added saving on changes
