+++ src/elements/row.js
@@ -1,5 +1,6 @@
 import createChild from 'elements/child';
 import { createJson } from 'index';
+import { controlsBinding, globalControls, dragNdrop  } from 'elements/events';
 
 export function makeId(){
         let text = 'group_';
@@ -8,31 +9,12 @@ export function makeId(){
         return text;
     };
 
-export function dragNdrop(){
-     $('main .container').shapeshift({
-        selector: 'div',
-        minColumns: 3,
-        colWidth: 235,
-        align: 'left',
-     });
-    $('aside .tabs').shapeshift({
-        selector: 'div',
-        minColumns: 1,
-         maxColumns: 1,
-        colWidth: 300
-    });
-     $('.container').on( 'ss-added', function(e) {
-        chrome.storage.local.clear();
-        chrome.storage.local.set(createJson(), function() {
-            console.info('Settings saved');
-        });
-    });
-}
+
 
 export function createControls(title,id, isTabs) {
     const controls = $('<span>', { 'class': `controls` });
     const h1 = $('<h1 contenteditable="true" class="lead" data-type="title">'+title+'</h1>');
-    const close = $('<button>', { text:'Remove Row', 'class': 'danger','data-id':id });
+    const close = $('<button>', { text:'Remove Row', 'class': 'delete-row danger','data-id':id });
     controls.append(h1);
     if (isTabs !== 'tabs') {
         controls.append(close);
@@ -46,13 +28,10 @@ export function createRow(extraClass, title){
     const container =  $('<div>', { 'id': id, 'class': `container ${extraClass}`,'data-id':id });
     const groupTitle = title ? title : 'untitled'
     container.append(createControls(groupTitle, id, extraClass))
+    controlsBinding();
     return container
 }
-$('#createRow').on('click', function () {
-    const container = createRow('group',false);
-    $('main').append(container);
-    dragNdrop();
-});
+
 
 export function createTabs(){
     chrome.tabs.query({
@@ -72,7 +51,6 @@ export function createTabs(){
 
 export function createMarks(json){
     // load(function (json) {
-    console.log('json',json);
         Object.keys(json).map(function(key, index) {
             const group = json[key];
             var container = createRow('group',group.name);
@@ -84,7 +62,6 @@ export function createMarks(json){
                     });
                 }
             $('main').append(container);
+            controlsBinding();       
         });
-    // });
-
 }
